CB-30447: Disable mid-burning reboot in case of ARM64 + RHEL9

+++ scripts/reboot.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
-echo "Rebooting before validation and cleanup"
-reboot
+# This reboot causes problems with RHEL 9 ARM64 images and right now, not sure why...
+# We need to see if we need it at all. For now, it is disabled for this combo, but further testing is required.
+if [[ "${OS}" != "redhat9" && "${ARCHITECTURE}" != "arm64" ]] ; then
+  echo "Rebooting before validation and cleanup"
+  reboot
+fi
